snapcraft/cmds.py: ignore symlinks so that the python3 symlink is not replaced with a python3.real empty file

+++ snapcraft/cmds.py
@@ -75,6 +75,9 @@ def assemble(args):
             if exe.endswith('.real'):
                 continue
             exePath = os.path.join(absbindir, exe)
+            # ignore symlinks (e.g. python3 -> python3.4)
+            if os.path.islink(exePath):
+                continue
             try:
                 os.remove(exePath + '.real')
             except:
